# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Causal desk",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/datasets": {
      "get": {
        "summary": "List Datasets",
        "operationId": "list_datasets_api_datasets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetList"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Dataset",
        "operationId": "create_dataset_api_datasets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/profile": {
      "post": {
        "summary": "Profile Upload",
        "operationId": "profile_upload_api_profile_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_profile_upload_api_profile_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/datasets/{name}": {
      "delete": {
        "summary": "Delete Dataset",
        "operationId": "delete_dataset_api_datasets__name__delete",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{name}": {
      "get": {
        "summary": "Get Session",
        "operationId": "get_session_api_sessions__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{name}/messages": {
      "post": {
        "summary": "Post Message",
        "operationId": "post_message_api_sessions__name__messages_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageIn"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{name}/resume": {
      "post": {
        "summary": "Resume Session",
        "operationId": "resume_session_api_sessions__name__resume_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{name}/analyses": {
      "post": {
        "summary": "New Analysis",
        "description": "A new question on the same file: a new thread over the memory as it stands; every run so far stays listed.",
        "operationId": "new_analysis_api_sessions__name__analyses_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "summary": "Run Files",
        "operationId": "run_files_api_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunFiles"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/files/{filename}": {
      "get": {
        "summary": "Run File",
        "operationId": "run_file_api_runs__run_id__files__filename__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          },
          {
            "name": "raw",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Raw"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/artifacts/{name}/{moment}/{design}/{artifact_id}/figure.png": {
      "get": {
        "summary": "Artifact Figure",
        "operationId": "artifact_figure_api_artifacts__name___moment___design___artifact_id__figure_png_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "moment",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Moment"
            }
          },
          {
            "name": "design",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Design"
            }
          },
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Activity": {
        "properties": {
          "node": {
            "type": "string",
            "title": "Node"
          },
          "since": {
            "type": "string",
            "title": "Since"
          }
        },
        "type": "object",
        "required": [
          "node",
          "since"
        ],
        "title": "Activity"
      },
...
      "Body_profile_upload_api_profile_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_profile_upload_api_profile_post"
      },
...
      "ColumnSummary": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "nulls": {
            "type": "integer",
            "title": "Nulls"
          },
          "null_rate": {
            "type": "number",
            "title": "Null Rate",
            "default": 0.0
          },
          "distinct": {
            "type": "integer",
            "title": "Distinct"
          },
          "constant": {
            "type": "boolean",
            "title": "Constant",
            "default": false
          },
          "examples": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Examples"
          },
          "numeric": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NumericShape"
              },
              {
                "type": "null"
              }
            ]
          },
          "top_values": {
            "items": {
              "$ref": "#/components/schemas/TopValue"
            },
            "type": "array",
            "title": "Top Values"
          },
          "datetime": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DatetimeShape"
              },
              {
                "type": "null"
              }
            ]
          },
          "sentinels": {
            "items": {
              "$ref": "#/components/schemas/Sentinel"
            },
            "type": "array",
            "title": "Sentinels"
          },
          "issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "name",
          "key",
          "kind",
          "nulls",
          "distinct",
          "examples"
        ],
        "title": "ColumnSummary",
        "description": "One column as the profiler read it. Facts only: shape, missingness, and anything odd."
      },
      "DatasetCreate": {
        "properties": {
          "name": {
            "type": "string",
            "pattern": "^[a-z][a-z0-9_]{1,39}$",
            "title": "Name"
          },
          "title": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Title"
          },
          "upload_id": {
            "type": "string",
            "pattern": "^[0-9a-f]{8}$",
            "title": "Upload Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "title",
          "upload_id"
        ],
        "title": "DatasetCreate",
        "description": "CSV only. The conversation starts with the causal question; nothing about the data is typed into a form."
      },
      "DatasetList": {
        "properties": {
          "datasets": {
            "items": {
              "$ref": "#/components/schemas/DatasetSummary"
            },
            "type": "array",
            "title": "Datasets"
          }
        },
        "type": "object",
        "required": [
          "datasets"
        ],
        "title": "DatasetList"
      },
      "DatasetSummary": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "csv": {
            "type": "string",
            "title": "Csv"
          },
          "rows": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows"
          },
          "columns": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Columns"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "shipped": {
            "type": "boolean",
            "title": "Shipped"
          },
          "has_claims": {
            "type": "boolean",
            "title": "Has Claims"
          },
          "question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          },
          "session": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SessionBrief"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "name",
          "title",
          "csv",
          "shipped",
          "has_claims"
        ],
        "title": "DatasetSummary"
      },
...
      "FileEntry": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          }
        },
        "type": "object",
        "required": [
          "name",
          "size"
        ],
        "title": "FileEntry"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageIn": {
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "MessageIn"
      },
...
      "ProfileOut": {
        "properties": {
          "upload_id": {
            "type": "string",
            "title": "Upload Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "rows": {
            "type": "integer",
            "title": "Rows"
          },
          "columns": {
            "items": {
              "$ref": "#/components/schemas/ColumnSummary"
            },
            "type": "array",
            "title": "Columns"
          },
          "head": {
            "items": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Head"
          },
          "duplicate_rows": {
            "type": "integer",
            "title": "Duplicate Rows",
            "default": 0
          },
          "candidate_keys": {
            "items": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Candidate Keys"
          },
          "grain": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grain"
          },
          "co_missing": {
            "items": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Co Missing"
          },
          "issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "upload_id",
          "filename",
          "rows",
          "columns"
        ],
        "title": "ProfileOut"
      },
      "Prompt": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "default": ""
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": ""
          },
          "ready": {
            "type": "boolean",
            "title": "Ready",
            "default": false
          },
          "open": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Open"
          },
          "runs": {
            "type": "integer",
            "title": "Runs",
            "default": 0
          },
          "phase": {
            "type": "string",
            "title": "Phase",
            "default": "before"
          },
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind",
            "description": "question: the causal question is being asked; ask: a field; after: the chat after a run"
          }
        },
        "type": "object",
        "title": "Prompt"
      },
...
      "RunFiles": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/FileEntry"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "files"
        ],
        "title": "RunFiles"
      },
...
      "SessionBrief": {
        "properties": {
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "phase": {
            "type": "string",
            "title": "Phase"
          },
          "runs": {
            "type": "integer",
            "title": "Runs"
          }
        },
        "type": "object",
        "required": [
          "stage",
          "phase",
          "runs"
        ],
        "title": "SessionBrief"
      },
      "SessionView": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          },
          "stage": {
            "type": "string",
            "enum": [
              "busy",
              "waiting",
              "ended",
              "stale",
              "error",
              "new"
            ],
            "title": "Stage"
          },
          "phase": {
            "type": "string",
            "title": "Phase",
            "default": "before"
          },
          "activity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Activity"
              },
              {
                "type": "null"
              }
            ]
          },
          "ready": {
            "type": "boolean",
            "title": "Ready",
            "default": false
          },
          "prompt": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Prompt"
              },
              {
                "type": "null"
              }
            ]
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionView"
            },
            "type": "array",
            "title": "Questions"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/ClaimView"
            },
            "type": "array",
            "title": "Claims"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StatusView"
              },
              {
                "type": "null"
              }
            ]
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunView"
            },
            "type": "array",
            "title": "Runs"
          },
          "brief": {
            "type": "string",
            "title": "Brief",
            "default": ""
          },
          "transcript": {
            "items": {
              "$ref": "#/components/schemas/Turn"
            },
            "type": "array",
            "title": "Transcript"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "analysis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis",
            "description": "a<k>: this conversation's journal"
          },
          "journal": {
            "items": {
              "$ref": "#/components/schemas/StepView"
            },
            "type": "array",
            "title": "Journal",
            "description": "the conversation's steps, in order"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactView"
            },
            "type": "array",
            "title": "Artifacts",
            "description": "the pictures drawn before any run, oldest first"
          }
        },
        "type": "object",
        "required": [
          "name",
          "title",
          "stage"
        ],
        "title": "SessionView"
      },
...
      "TopValue": {
        "properties": {
          "value": {
            "type": "string",
            "title": "Value"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "share": {
            "type": "number",
            "title": "Share"
          }
        },
        "type": "object",
        "required": [
          "value",
          "count",
          "share"
        ],
        "title": "TopValue"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}